+++ packages/flutter/lib/src/native/java/sentry_native_java.dart
@@ -23,6 +23,10 @@ class SentryNativeJava extends SentryNativeChannel {
   bool get supportsReplay => true;
 
   @override
+  SentryId? get replayId => _replayId;
+  SentryId? _replayId;
+
+  @override
   Future<void> init(Hub hub) async {
     // We only need these when replay is enabled (session or error capture)
     // so let's set it up conditionally. This allows Dart to trim the code.
@@ -30,14 +34,25 @@ class SentryNativeJava extends SentryNativeChannel {
       channel.setMethodCallHandler((call) async {
         switch (call.method) {
           case 'ReplayRecorder.start':
-            final replayId =
-                SentryId.fromId(call.arguments['replayId'] as String);
+            final replayIdArg = call.arguments['replayId'];
+            final replayIsBufferingArg = call.arguments['replayIsBuffering'];
+
+            final replayId = replayIdArg != null
+                ? SentryId.fromId(replayIdArg as String)
+                : null;
+
+            final replayIsBuffering = replayIsBufferingArg != null
+                ? replayIsBufferingArg as bool
+                : false;
+
+            _replayId = replayId;
 
             _replayRecorder = AndroidReplayRecorder.factory(options);
             await _replayRecorder!.start();
             hub.configureScope((s) {
+              // Only set replay ID on scope if not buffering (active session mode)
               // ignore: invalid_use_of_internal_member
-              s.replayId = replayId;
+              s.replayId = !replayIsBuffering ? replayId : null;
             });
             break;
           case 'ReplayRecorder.onConfigurationChanged':
@@ -78,6 +93,13 @@ class SentryNativeJava extends SentryNativeChannel {
     await _envelopeSender?.start();
 
     return super.init(hub);
+  }
+
+  @override
+  FutureOr<SentryId> captureReplay() async {
+    final replayId = await super.captureReplay();
+    _replayId = replayId;
+    return replayId;
   }
 
   @override